+++ src/main/java/io/botcrafting/botcraft/controller/message/MessageController.java
@@ -1,25 +1,19 @@
 package io.botcrafting.botcraft.controller.message;
 
 import io.botcrafting.botcraft.model.Update;
-import io.botcrafting.botcraft.model.service.MessageService;
+import io.botcrafting.botcraft.model.usecase.MessageUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class MessageController {
     @Autowired
-    private MessageService messageService;
+    private MessageUseCase messageUseCase;
 
     @RequestMapping(value = "/", method = RequestMethod.POST)
     @ResponseBody
     public String receiveTelegramUpdate(@RequestBody Update telegramUpdate){
-        messageService.handle(telegramUpdate);
-        return "OK";
-    }
-
-    @RequestMapping(value = "/keep-alive" , method = RequestMethod.POST)
-    @ResponseBody
-    public String receiveKeepAliveRequest(){
+        messageUseCase.handle(telegramUpdate);
         return "OK";
     }
 }
